requirements.txt
beautifulsoup4==4.12.2
python-wordpress-xmlrpc==2.3
python-telegram-bot==20.7
selenium==4.16.0
webdriver-manager==4.0.1
PyYAML==6.0.1

beautifulsoup4==4.12.2
python-wordpress-xmlrpc==2.3
python-telegram-bot==20.7
selenium==4.16.0
webdriver-manager==4.0.1
PyYAML==6.0.1

gpt4all==2.0.2